!hotfix 브랜치를 develop으로 머지

+++ TimeTableArtist/src/main/java/SamwaMoney/TimeTableArtist/Reply/service/ReplyService.java
@@ -6,7 +6,6 @@ import SamwaMoney.TimeTableArtist.Reply.domain.Reply;
 import SamwaMoney.TimeTableArtist.Reply.dto.ReplyRequestDto;
 import SamwaMoney.TimeTableArtist.Reply.repository.ReplyRepository;
 import SamwaMoney.TimeTableArtist.Timetable.domain.Timetable;
-import SamwaMoney.TimeTableArtist.Timetable.service.AllClassAlgoService;
 import SamwaMoney.TimeTableArtist.Timetable.service.TimetableService;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Transactional
@@ -34,14 +34,9 @@ public class ReplyService {
         boolean nameHide = requestDto.isNameHide();
         String name = writer.getUsername();
         String replyName = selectName(nameHide, name);
-        Long replyCount = timetable.getReplyCount();
-
         boolean isHeart = false;
         Integer replyLikeCount = 0;
-
-        Long newReplyCount = replyCount + 1;
-        timetable.updateReplyCount(newReplyCount);
-
+        timetable.setReplyCount(timetable.getReplyCount() + 1);
         return replyRepository.save(requestDto.toEntity(timetable, writer, replyName, isHeart, replyLikeCount)).getReplyId();
     }
 
@@ -67,18 +62,17 @@ public class ReplyService {
                 .orElseThrow(()->new EntityNotFoundException("해당 댓글이 존재하지 않습니다."));
     }
 
-    public void removeReply(Long replyId){
+    public String removeReply(Long memberId, Long replyId){
         Reply reply = findReplyById(replyId);
-        replyRepository.delete(reply);
-        Timetable timetable = reply.getTimetable();
-        Integer replyCount = getReplyCount(timetable);
-        timetable.updateReplyCount(Long.valueOf(replyCount));
-    }
-
-    @Transactional (readOnly = true)
-    public Integer getReplyCount(Timetable timetable) {
-        Integer replyCount = replyRepository.countByTimetable(timetable);
-        return replyCount;
+        if (Objects.equals(reply.getWriter().getMemberId(), memberId)){
+            Timetable timetable = findReplyById(replyId).getTimetable();
+            timetable.setReplyCount(timetable.getReplyCount()-1);
+            replyRepository.delete(reply);
+            return "댓글이 삭제되었습니다.";
+        }
+        else{
+            return "해당 댓글을 삭제할 권한이 없습니다.";
+        }
     }
 }
